Improve XML integer decoding.
In the XML decoder, ints and int64s are constrained when they overflow.

Also worth noting is that, when invalid characters are parsed for an int
element, Roblox discards the element, causing the default value to be
used rather than setting the value to zero. This was verified by using
the Debris.MaxItems property, which is an int with a non-zero default
value of 1000. It is assumed that int64 has the same behavior.