Improve support for typeless PROP chunks.
In at least one version of Roblox, Studio produced file with the
Workspace.WorldPivotInternal property as a chunk with no type or
values. That is, the chunk simply ended after the property name.

Roblox has since removed the property, and seems to handle such chunks
by silently ignoring them. This changes improves support for such
chunks by having the decoder emit a warning when it encounters a chunk,
then discarding it. The encoder never produces such chunks.