Fix rbxl implementation of Optional.
rbxfile.ValueOptional refactored to always have a type. These are
enforced by the new Some and None constructors.

Refactor out valueOptional type. Implementation of optional is
customized enough that valueOptional is unnecessary. All functionality
is handled by arrayOptional. Nil is used arrayOptional to indicate the
lack of value.

arrayOptional must be initialized with a type before Get and Set can
be used.

Notably, an optional of type optional is not handled correctly. It may
also be possible that other types with custom handling are not handled
correctly either. These may be resolved by disallowing certain types as
optionals.