Fix UniqueId chunks not being removed from models.
If the program encodes a UniqueId type while not in Place mode, then it
is supposed to skip that type. It does this, but does not remove the
chunk. This produces a chunk that does not have its Properties field
set. Code later in the program assumes that this field is set,
resulting in a nil pointer dereference.

Resolved by removing the chunk.